requirements.txt
-i https://pypi.org/simple
attrs==23.1.0; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
click==8.1.6; python_version >= '3.7'
flask==2.2.5
flask-cors==4.0.0
importlib-metadata==6.7.0; python_version < '3.10'
importlib-resources==5.12.0; python_version < '3.9'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
jsonschema==4.17.3
markupsafe==2.1.3; python_version >= '3.7'
mysql-connector-python==8.0.33
pkgutil-resolve-name==1.3.10; python_version < '3.9'
protobuf==3.20.3; python_version >= '3.7'
psycopg==3.1.10
pyrsistent==0.19.3; python_version >= '3.7'
python-dotenv==0.21.1
typing-extensions==4.7.1; python_version < '3.8'
werkzeug==2.2.3; python_version >= '3.7'
zipp==3.15.0; python_version >= '3.7'
